rename NodeManager for ParserManager
