+++ src/orm/manager/index.ts
@@ -1,4 +1,3 @@
 export * from './expression'
-export * from './nodeExpression'
 export * from './memoryCache'
 export * from './transaction'
