+++ src/orm/parser/index.ts
@@ -1,3 +1,3 @@
 export * from './model'
 export * from './node'
-export * from './nodeManager'
+export * from './parserManager'
